Push query methods to BSONStore
